refactor(codex): split lineage responsibilities
